feat: integrate ServiceNetworkDoctor for network diagnostics and update NetworkAlert component
