+++ packages/kit-bg/src/apis/IBackgroundApi.ts
@@ -58,6 +58,7 @@ import type ServiceMarketWS from '../services/ServiceMarketWS';
 import type ServiceMasterPassword from '../services/ServiceMasterPassword';
 import type ServiceNameResolver from '../services/ServiceNameResolver';
 import type ServiceNetwork from '../services/ServiceNetwork';
+import type ServiceNetworkDoctor from '../services/ServiceNetworkDoctor';
 import type ServiceNFT from '../services/ServiceNFT';
 import type ServiceNostr from '../services/ServiceNostr';
 import type ServiceNotification from '../services/ServiceNotification';
@@ -224,4 +225,5 @@ export interface IBackgroundApi extends IBackgroundApiBridge {
   serviceHyperliquidSubscription: ServiceHyperliquidSubscription;
 
   serviceIpTable: ServiceIpTable;
+  serviceNetworkDoctor: ServiceNetworkDoctor;
 }
